+++ vamp-plugins/dsp/PhaseVocoder.cpp
@@ -25,8 +25,8 @@ PhaseVocoder::PhaseVocoder(unsigned int n) :
     m_n(n)
 {
     m_fft = new FFTReal(m_n);
-    m_realOut = new double[m_n];
-    m_imagOut = new double[m_n];
+    m_realOut = new float[m_n];
+    m_imagOut = new float[m_n];
 }
 
 PhaseVocoder::~PhaseVocoder()
@@ -36,17 +36,17 @@ PhaseVocoder::~PhaseVocoder()
     delete m_fft;
 }
 
-void PhaseVocoder::FFTShift(unsigned int size, double *src)
+void PhaseVocoder::FFTShift(unsigned int size, float *src)
 {
     const int hs = size/2;
     for (int i = 0; i < hs; ++i) {
-        double tmp = src[i];
+        float tmp = src[i];
         src[i] = src[i + hs];
         src[i + hs] = tmp;
     }
 }
 
-void PhaseVocoder::process(double *src, double *mag, double *theta)
+void PhaseVocoder::process(float *src, float *mag, float *theta)
 {
     FFTShift( m_n, src);
 	
@@ -56,7 +56,7 @@ void PhaseVocoder::process(double *src, double *mag, double *theta)
     getPhase( m_n/2, theta, m_realOut, m_imagOut);
 }
 
-void PhaseVocoder::getMagnitude(unsigned int size, double *mag, double *real, double *imag)
+void PhaseVocoder::getMagnitude(unsigned int size, float *mag, float *real, float *imag)
 {	
     unsigned int j;
 
@@ -66,7 +66,7 @@ void PhaseVocoder::getMagnitude(unsigned int size, double *mag, double *real, do
     }
 }
 
-void PhaseVocoder::getPhase(unsigned int size, double *theta, double *real, double *imag)
+void PhaseVocoder::getPhase(unsigned int size, float *theta, float *real, float *imag)
 {
     unsigned int k;
